requirements.txt
python==3.7
tensorflow-gpu==2.3.1
soundfile
torchaudio==0.5
git+https://github.com/pvigier/perlin-numpy
librosa
matplotlib
tqdm